% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/measure_ent.R
\name{compute_ent}
\alias{compute_ent}
\title{compute_ent Function}
\usage{
compute_ent(xx)
}
\arguments{
\item{xx}{a table of (joint) probabilities}
}
\description{
This function computes entropy
}
\examples{
compute_ent ()
}
\keyword{entropy}

