% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/measure_disc.R
\name{measure_disc}
\alias{measure_disc}
\title{measure_disc Function}
\usage{
measure_disc(data)
}
\arguments{
\item{data}{data containint at least a column 'y' (target) and a column 's' (protected attribute)}
}
\description{
This function allows you to masure discrimination in data using different measures.
}
\examples{
measure_disc ()
}
\keyword{discrimination}
\keyword{measures}

