% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salary}
\alias{salary}
\title{salary}
\format{A data frame with 52 observations and five variables:
\describe{
\item{\code{sex}}{gender: 1 - female, 0 - male}
\item{\code{rank}}{professor rank: 1 - assistant, 2 - associate, 3 - full}
\item{\code{degree}}{degree: 0 - masters, 1 - doctorate}
\item{\code{yeardegree}}{years since degree}
\item{\code{salary}}{salary in USD for 9 months}
}

Obtained from \url{http://data.princeton.edu/wws509/datasets/\#salary}}
\usage{
salary
}
\description{
Salaries of professors at a US college.
}
\keyword{datasets}

