% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyr}
\alias{toyr}
\title{toyr}
\format{A data frame with 10 observations and three variables:
\describe{
\item{\code{education}}{presumaby in years}
\item{\code{ethnicity}}{presumably 1 means foreigner (deprived), 0 - native}
\item{\code{salary}}{presumed monthly salary}
}}
\usage{
toyr
}
\description{
A toy dataset for regression generated as sal = 1000 + 100*edu - 500*etn
}
\keyword{datasets}

